Move external calls commands to integrations
