+++ anton/commands/ui.py
@@ -52,6 +52,7 @@ def print_slash_help(console: Console) -> None:
     console.print("  [bold]/paste[/]     — Attach an image from your clipboard")
     console.print("  [bold]/resume[/]    — Continue a previous session")
     console.print("  [bold]/publish[/]   — Publish an HTML report to the web")
+    console.print("  [bold]/unpublish[/] — Remove a published report")
 
     console.print("\n[bold]General[/]")
     console.print("  [bold]/help[/]      — Show this help menu")
